refactor(ms2): component for password input and confirmation
